+++ src/main/scala/satify/dsl/Numbers.scala
@@ -1,7 +1,6 @@
 package satify.dsl
 
 object Numbers:
-
   val one = 1
   val two = 2
   val three = 3
